requirements.txt
accelerate==1.12.0
aiofiles==24.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
brotli==1.2.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
diffusers==0.36.0
exceptiongroup==1.3.1
fastapi==0.128.0
ffmpy==1.0.0
filelock==3.20.0
fonttools==4.61.1
fsspec==2025.12.0
gradio==6.2.0
gradio_client==2.0.2
groovy==0.1.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
hypothesis==6.148.8
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
inquirerpy==0.3.4
Jinja2==3.1.6
kiwisolver==1.4.9
lpips==0.1.4
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
orjson==3.11.5
packaging==25.0
pandas==2.3.3
peft==0.18.0
pfzy==0.3.4
pillow==12.0.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psutil==7.2.0
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.3.1
pytest==9.0.2
python-dateutil==2.9.0.post0
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rich==14.2.0
safehttpx==0.1.7
safetensors==0.7.0
scipy==1.15.3
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sortedcontainers==2.4.0
starlette==0.50.0
sympy==1.14.0
tokenizers==0.22.1
tomli==2.3.0
tomlkit==0.13.3
torch==2.9.1+cu128
torchaudio==2.9.1+cu128
torchvision==0.24.1+cu128
tqdm==4.67.1
transformers==4.57.3
typer==0.21.0
typer-slim==0.21.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.40.0
wcwidth==0.2.14
zipp==3.23.0
